% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_cleaning-functions.R
\name{chunk_it}
\alias{chunk_it}
\title{Break up a thing into chunks of a certain size n
break up into chunks of size n}
\usage{
chunk_it(x, n)
}
\description{
Break up a thing into chunks of a certain size n
break up into chunks of size n
}
