% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_era5.R
\name{clean_era5}
\alias{clean_era5}
\title{Clean ERA5 Rasters}
\usage{
clean_era5(
  input_path,
  input_filename = "example_era5",
  input_filetype = "nc",
  keep_substrings = c("skt"),
  output_path,
  output_filename = NULL
)
}
\arguments{
\item{input_path}{file path to your raw ERA data (e.g. file.path("E:","Projects","data","01_raw",ERA_5"))}

\item{input_filename}{the short name that you use to distinguish the .nc file}

\item{input_filetype}{the type of input file; needs to be something terra::rast() can read}

\item{keep_substrings}{a character vector of the substrings which define which layers you would like to keep}

\item{output_path}{the file path to put your cleaned data. Creates it if it doesn't already exist, e.g. file.path("E:","Projects","data","03_clean","ERA_5")}

\item{output_filename}{the name to give the cleaned ERA5 raster. NULL defaults to:
paste0(input_filename,".tif"); note this uses terra::writeRaster, not saveRDS
@return just puts the correctly rotated and subsetted raster into the output_path folder}
}
\description{
takes as data a raster file (ideally taken from Copernicus ERA5),
selects a particular variable to analyze and subsets the rasters according to
this variable, and rotates the rasters so they can be extracted to more typical polygon data
}
\examples{
clean_era5(input_path      = file.path(data_external_raw,"ERA_5"),
input_filename  = "example_era5",
input_filetype  = "nc",
keep_substrings = c("skt"),
output_path     = file.path(data_external_clean,"ERA_5"),
output_filename = paste0("example_era5.tif"))


}
\author{
jillianstallman
}
