% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{create_long_run_vars}
\alias{create_long_run_vars}
\title{Description: Create Long-run variables}
\usage{
create_long_run_vars(
  panel_df,
  id_varname = "DHSID",
  time_varname = "year",
  variable_to_manipulate = "precip_annual_mean",
  variable_namestub = "precip",
  out_path = NULL,
  out_filename = NULL
)
}
\arguments{
\item{panel_df}{panel data frame}

\item{id_varname}{the variable name of the}

\item{time_varname}{variable name of the largest}

\item{variable_to_manipulate}{character vector of the variable that you want long-run values for}

\item{variable_namestub}{character vector, the namestub that you want appended to the long-run varnames}

\item{out_path}{output file.path() to where to save the final data frame. Default is NULL, in which case the df isn't saved}

\item{output_filename}{filename to save the final data frame, if you provide an out_path. Default NULL doesn't do anything}
}
\value{
a data frame that for the variable_to_manipulate
}
\description{
With a df with a time dimension, create a bunch of long-run variables
}
\examples{
# example code

}
