% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_download-functions.R
\name{decompress_file}
\alias{decompress_file}
\title{function to decompress files that are large
https://stackoverflow.com/questions/42740206/r-possible-truncation-of-4gb-file}
\usage{
decompress_file(directory, file, .file_cache = FALSE)
}
\description{
function to decompress files that are large
https://stackoverflow.com/questions/42740206/r-possible-truncation-of-4gb-file
}
