% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_plot-functions.R
\name{density_plot_all_layers}
\alias{density_plot_all_layers}
\title{Generates density plots of a base dataframe in one color and another data list
that has the same sort of data each as elements of that list, e.g. for
plotting a regular observed outcome and bootstrapped outcomes}
\usage{
density_plot_all_layers(
  base_df,
  plot_var_index,
  layers_list,
  where_text_y,
  where_text_x,
  ...
)
}
\description{
Generates density plots of a base dataframe in one color and another data list
that has the same sort of data each as elements of that list, e.g. for
plotting a regular observed outcome and bootstrapped outcomes
}
