% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_era5.R
\name{download_era5}
\alias{download_era5}
\title{Downloads ERA5 data using R and python}
\usage{
download_era5(
  raw_data_path,
  code_download_path,
  cdsapi_filename,
  new_era5_filename = NULL
)
}
\arguments{
\item{raw_data_path}{path location where you want to download the ERA5 data to. the data will be downloaded into a file called file.path(raw_data_path,"ERA_5") which will be created if it doesn't exist already}

\item{cdsapi_filename}{the filename with ".py" extension that you labeled the API call as}

\item{new_era5_filename}{the default CDS API filename is "download.nc". If you did not change this default and have multiple API calls this will lead to conflicts, so set new_era5_filename as "newname.nc" to avoid your file getting overwritten}

\item{cdsapi_filepath}{path location of the .py file you downloaded from Copernicus}
}
\description{
\code{download_era5} use the ERA5 API to download desired data
See the README associated with the era5-extraction repository for instrucitons on how to get the CDS API request file and be able to download the ERA5 data using the Copernicus API
If you are getting the error from Python that says something like "cdsapi module not found', it is likely that r-reticulate is running into an issue finding your python virtual environment and you are running this code within a project. To trouble-shoot, try going to Tools -> Global Options -> Python and uncheck "Automatically activate project-local Python environments." See https://github.com/rstudio/reticulate/issues/1362 for additional troubleshooting.
}
\author{
jillianstallman
}
