% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-functions.R
\name{get_parallel_splits}
\alias{get_parallel_splits}
\title{Split an object to send these split objects to n_cores number of workers}
\usage{
get_parallel_splits(thing_to_split, n_cores)
}
\arguments{
\item{thing_to_split}{is the df, vector or list to get subsetted}

\item{n_cores}{integer, the number of cores (and thus the number of elements of the list)}
}
\value{
split_list a list with n_cores elements that splits up thing_to_split into n_cores elements
}
\description{
Takes a thing (tibble, df, whatever) and splits it into n_cores chunks to be put into a list
The list can then be the input into a parLapply or other parallel function
}
