% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_plot-functions.R
\name{get_table}
\alias{get_table}
\title{Share plots
get a frequency table, which will be a good input into the following
share_plot function}
\usage{
get_table(data_call)
}
\description{
Share plots
get a frequency table, which will be a good input into the following
share_plot function
}
