% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_plot-functions.R
\name{make_histogram}
\alias{make_histogram}
\title{Function make_histogram: generates a histogram with labels and vertical lines}
\usage{
make_histogram(
  data_frame,
  counting_var_index,
  title,
  caption,
  where_y = c(7000, 8200, 2000, 2000),
  where_x = c(0.98, 0.7, 0.02, 0.85),
  fill_color = "#63aaff",
  text_color = "#00356b",
  ...
)
}
\arguments{
\item{data_frame}{the data frome you want to make a histogram in}

\item{counting_var_index}{numeric, the column that you want to take the hist over}

\item{title}{character, title of graph}

\item{caption}{character, caption of graph}

\item{where_y}{vector, contains 4 numerics with the y placement of the text labels}

\item{where_x}{vector, contains 4 numerical vars, where to place the text for
the 95th percentile, median, 5th percentile, and mean, in terms of where the percentile
distribution is, e.g. c(.98,.70,.02,.85),}

\item{...}{goes into theme plot}

\item{barcolor}{the color of the histogram}
}
\value{
returns the plot object of your histogram
}
\description{
Function make_histogram: generates a histogram with labels and vertical lines
}
