% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_extract_to_long_df.R
\name{raster_extract_to_long_df}
\alias{raster_extract_to_long_df}
\title{Extracts terra rasters to a vector SF}
\usage{
raster_extract_to_long_df(
  terra_raster,
  vector_sf,
  cast_vector_sf = FALSE,
  vector_cast_out_path = getwd(),
  vector_cast_out_filename = "vector_cast.rds",
  save_raster_copy = FALSE,
  raster_out_path = getwd(),
  raster_out_filename = "terra_raster_rotated_and_wgs84.tif",
  extracted_out_path = getwd(),
  extracted_out_filename = "terra_raster_extracted_to_vector.rds",
  layer_substrings = "all",
  layer_names_vec = NULL,
  layer_names_title = "date",
  func = "mean",
  weights = NULL,
  remove_files = FALSE,
  ...
)
}
\arguments{
\item{terra_raster}{a terra raster. If not from 0-360 degrees longitude and -90 to 90 degrees latitude will rotate the latitude (which takes a while). Easiest to use the output of clean_era5 to generate this input raster.}

\item{vector_sf}{an sf object with polygons, points or lines to be extracted to}

\item{cast_vector_sf}{logical with default FALSE. set this to TRUE if you have a multipolygon,multipoint or multilinestring and you want to create a sf in which each row gives a single geometry unit rather than aggregating over possibly geographically disparate units (e.g. if you are aggregating temperature up to a mean annual value for a country and are using the United States, this includes Alaska and Hawaii which geographically are located far from the contiguous continental United States, which may not give you the average you wanted)}

\item{vector_cast_out_path}{if casting multipolygons to their component polygons, filepath to where the vector sf that gets cast is output. Default is working directory}

\item{vector_cast_out_filename}{if cast_vector_sf is TRUE, filename for the vector sf output. default is vector_cast.rds}

\item{save_raster_copy}{default FALSE, logical. Set to TRUE if you want to save a copy of the raster you bring in
with possibly updated coordinate reference system or rotated correctly}

\item{raster_out_path}{if you're saving a copy of the raster, file path to save your copy of the raster to. Default is working directory.}

\item{raster_out_filename}{if saving a raster copy, name of the file you want to save it as. Default is "terra_raster_rotated_and_wgs84.tif"}

\item{extracted_out_path}{filepath, where you want to save the vector sf with its extracted values}

\item{extracted_out_filename}{filename of output file from extracted raster to vector (no geometry kept)}

\item{layer_substrings}{character vector, giving the substrings to identify layers to keep. Default is "all," keep everything. But if you have a single layer type (e.g. surface temperature), change this because it'll become part of the column name for the long data frame.}

\item{layer_names_vec}{character vector, needs to be of the same length as the number of layers of your raster. Default NULL populates this with the output of time(terra_raster)}

\item{layer_names_title}{default "date", the column(i.e. variable) name that you want for the ultimate data frame that has a column with title "layer_names_title" and which has the value for each raster layer the corresponding element of layer_names_vec.}

\item{func}{character vector, default "mean", also e.g. "weighted_sum" for exactextractr, function to be used for aggregating, from terra (points) or exactextractr (polygons)}

\item{weights}{character vector, default NULL. To be used for weighting the function, see packages terra (points) or exactextractr (polygons) for options, e.g. "area"}

\item{...}{additional options to be passed to terra::extract() (for points) or exactextractr::exact_extract() (for polygons/lines)}
}
\value{
a pivoted-long sf
}
\description{
\code{raster_extract_to_long_df} takes a terra spatial raster stack and extracts from this to either points, polygons or lines to generate for each spatial unit an extracted value for each raster layer. Outputs a long data frame where an observation is spatial-unit-by-raster-layer. This is typically going to be a spatial unit (e.g. city) by time (e.g. 2004 July) for a particular variable (e.g. sea surface temperature). Optionally casts polygon shapefile to one which is cast so that no single row contains multiple geometries.
}
\examples{
# example code
}
\author{
jillianstallman
}
