% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_extract_to_panel.R
\name{raster_extract_to_panel}
\alias{raster_extract_to_panel}
\title{Extracts terra rasters to a vector SF}
\usage{
raster_extract_to_panel(
  terra_raster,
  vector_sf,
  cast_vector_sf = FALSE,
  save_raster_copy = FALSE,
  raster_out_path = getwd(),
  raster_out_filename = "terra_raster.rds",
  extracted_out_path = getwd(),
  extracted_out_filename = "terra_raster_extracted_to_vector.rds",
  vector_cast_out_path = getwd(),
  vector_cast_out_filename = "vector_cast.rds",
  layer_substrings,
  func = "mean",
  weights = NULL,
  time_interval = "months",
  remove_files = FALSE
)
}
\arguments{
\item{terra_raster}{a terra raster, currently required to be from 0-360 degrees longitude and -90 to 90 degrees latitude, must have a time dimension given. you could use the clean_era5 function to create this raster, for example}

\item{vector_sf}{an sf object with polygons, points or lines to be extracted to}

\item{cast_vector_sf}{logical with default FALSE. set this to TRUE if you have a multipolygon,multipoint or multilinestring and you want to create a sf in which each row gives a single geometry unit rather than aggregating over possibly geographically disparate units (e.g. if you are aggregating temperature up to a mean annual value for a country and are using the United States, this includes Alaska and Hawaii which geographically are located far from the contiguous continental United States, which may not give you the average you wanted)}

\item{save_raster_copy}{default FALSE, logical. Set to TRUE if you want to save a copy of the raster you bring in 
with possibly updated coordinate reference system or rotated correctly}

\item{raster_out_path}{if you're saving a copy of the raster, file path to save your copy of the raster to}

\item{raster_out_filename}{if saving a raster copy, name of the file you want to save it as}

\item{extracted_out_path}{filepath, where you want to save the vector sf with its extracted values}

\item{extracted_out_filename}{filename of output file from extracted raster to vector (no geometry kept)}

\item{vector_cast_out_path}{if casting multipolygons to their component polygons, filepath to where the vector sf that gets cast is output}

\item{vector_cast_out_filename}{if cast_vector_sf is TRUE, filename for the vector sf output}

\item{layer_substrings}{character vector, giving the substrings to identify layers to keep.
assumes each element of this will give layers which generate the same longitudinal time sequence}

\item{func}{character vector, default "weighted_sum," function to be used for aggregating, from terra (points) or exactextractr (polygons)}

\item{weights}{character vector, default "area" to be used for weighting the function, see packages terra (points) or exactextractr (polygons) for options}
}
\description{
\code{raster_extract_to_panel} returns polygon shapefiles with longitudinal data from rasters
extracted to those polygons and writes two sf files, one with the polygons shapefiles, which are 
cast so that no single row contains multiple geometries,
and one with a long data frame that includes the raster extracted values 
and an identifier that allows for linking to the sf polygons.
}
\author{
jillianstallman
}
