% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_cleaning-functions.R
\name{save_rds_csv}
\alias{save_rds_csv}
\title{Save RDS and CSV ----}
\usage{
save_rds_csv(
  data,
  output_path,
  date = "",
  output_filename,
  remove = TRUE,
  csv_vars = c("all"),
  format = "both"
)
}
\arguments{
\item{data}{the data frame}

\item{output_path}{file path to the directory you want to store}

\item{date}{in character the dates relevant to the filename, will be put at the front of the filename}

\item{output_filename}{character, in RDS the filename output, assumes it ends in ".rds" and starts with "_" e.g. "_gkg_events.rds" so if date = "2016"
then the file would be called "2016_gkg_events.rds"}

\item{remove}{defaults to TRUE in which case the data are removed after being saved, if FALSE returns the data to memory}

\item{csv_vars}{vector of character strings with the varnames of the variables that will be saved in the CSV file}

\item{format}{defaults to "both" which is both csv and xlsx. Otherwise can use "csv" or "xlsx" for output format}
}
\description{
Given data, saves an output in RDS and CSV
}
