% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{share_plot_bottom}
\alias{share_plot_bottom}
\title{make a vertical share plot with the labels and percentages on it}
\usage{
share_plot_bottom(data_frame, fillvar, title, ylabel = "", ...)
}
\arguments{
\item{data_frame}{is the main data, can be output of get_table}

\item{fillvar}{write as data_frame$fillvar, the variable to put in shares.
unless otherwise manipulated, this is data_frame$Var1 if using the output of get_table}

\item{title}{plot title
... use to adjust the theme}
}
\description{
make a vertical share plot with the labels and percentages on it
}
