% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_map-functions.R
\name{theme_map}
\alias{theme_map}
\title{Set default theme for maps}
\usage{
theme_map(
  legend_text_size = 8,
  legend_title_size = 10,
  legend_position = c(0.2, 0.3),
  axis_title_x = element_text(color = "black"),
  axis_title_y = element_text(color = "black"),
  axis_text_x = element_text(color = "darkgrey"),
  axis_text_y = element_text(color = "darkgrey"),
  ...
)
}
\description{
Set default theme for maps
}
