% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_map-functions.R
\name{theme_map_gif}
\alias{theme_map_gif}
\title{set theme for maps when we have gifs}
\usage{
theme_map_gif(legend_text_size = 17, legend_title_size = 20)
}
\description{
set theme for maps when we have gifs
}
