% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_plot-functions.R
\name{theme_plot}
\alias{theme_plot}
\title{set a basic theme for plots}
\usage{
theme_plot(
  title_size = 25,
  axis_title_x = element_text(color = "black"),
  axis_title_y = element_text(color = "black"),
  axis_text_x = element_text(color = "darkgrey"),
  axis_text_y = element_text(color = "darkgrey"),
  ...
)
}
\description{
set a basic theme for plots
}
