% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{chunk_it}
\alias{chunk_it}
\title{Break up a thing into chunks of a certain size n}
\usage{
chunk_it(x, n)
}
\arguments{
\item{x}{a vector or list, things you want to split up}

\item{n}{number of chunks}
}
\description{
Break up a thing into chunks of a certain size n
}
\examples{
chunk_urls <- chunk_it(my_URLS,
                       n = ceiling(length(my_URLS)/22000))
missing_urls <- rep(NA, length(my_URLS))

missing_url_chunks <- chunk_it(missing_urls,
                               n = ceiling(length(my_URLS)/22000))
}
