% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_panel_to_annual.R
\name{collapse_panel_to_annual}
\alias{collapse_panel_to_annual}
\title{Collapse data to annual according to an id}
\usage{
collapse_panel_to_annual(
  in_df,
  date_varname = "date",
  value_varname_old = "precip_mean",
  value_varname_new = "precip",
  rolling_average_years = c(3, 5, 11),
  id_varname = "vector_cast_id",
  out_path = getwd(),
  out_filename = "annualized_df.rds"
)
}
\arguments{
\item{date_varname}{the variable name (as a character) that dates are stored as; needs to include year and month and be readable with as.Date()}

\item{value_varname_old}{the current name of the values you're using for the values column}

\item{value_varname_new}{new variable name for the values you're interested in summarizing from monthly to annually}

\item{id_varname}{the observation ID. default is "vector_cast_id" because that's what raster_extract_to_long_df outputs as the id, the observation ID}

\item{out_path}{the file.path() to where you would like to save the annual df. Gets created if it doesn't already exist. Defaults to working directory}

\item{out_filename}{The name you would like the long annual panel to be saved as. defaults to "annualized_df.rds"}

\item{panel_df}{the panel data frame which is currently monthly and we want to be annual}
}
\value{
an annualized df with 3 annual rolling averages
}
\description{
from a data frame that is higher-than-annual frequency, collapses to annual with rolling averages at 3, 5 and 11 years
}
\examples{
collapse_panel_to_annual(in_df = out_df, value_varname_old = "precip_mean",  value_varname_new = "precip",  id_varname = "vector_cast_id", out_path = file.path(data_external_clean,"merged","DHS_ERA5","annual"),  out_filename = "africa_dhs_gps_era5_annual.rds")
}
